+++ node/src/main/kotlin/net/corda/node/services/transactions/NonValidatingNotaryFlow.kt
@@ -5,10 +5,10 @@ import net.corda.core.identity.Party
 import net.corda.core.node.services.TrustedAuthorityNotaryService
 import net.corda.core.transactions.FilteredTransaction
 import net.corda.core.utilities.unwrap
-import net.corda.flows.NotaryFlow
+import net.corda.flows.NotarisationHandler
 import net.corda.flows.TransactionParts
 
-class NonValidatingNotaryFlow(otherSide: Party, service: TrustedAuthorityNotaryService) : NotaryFlow.Service(otherSide, service) {
+class NonValidatingNotaryFlow(otherSide: Party, service: TrustedAuthorityNotaryService) : NotarisationHandler(otherSide, service) {
     /**
      * The received transaction is not checked for contract-validity, as that would require fully
      * resolving it into a [TransactionForVerification], for which the caller would have to reveal the whole transaction
