+++ node/src/main/kotlin/net/corda/node/services/transactions/RaftValidatingNotaryService.kt
@@ -3,7 +3,7 @@ package net.corda.node.services.transactions
 import net.corda.core.identity.Party
 import net.corda.core.node.services.TrustedAuthorityNotaryService
 import net.corda.core.node.services.TimeWindowChecker
-import net.corda.flows.NotaryFlow
+import net.corda.flows.NotarisationHandler
 import net.corda.node.services.api.ServiceHubInternal
 
 /** A validating notary service operated by a group of mutually trusting parties, uses the Raft algorithm to achieve consensus. */
@@ -15,7 +15,7 @@ class RaftValidatingNotaryService(override val services: ServiceHubInternal) : T
     override val timeWindowChecker: TimeWindowChecker = TimeWindowChecker(services.clock)
     override val uniquenessProvider: RaftUniquenessProvider = RaftUniquenessProvider(services)
 
-    override fun createServiceFlow(otherParty: Party, platformVersion: Int): NotaryFlow.Service {
+    override fun createHandler(otherParty: Party, platformVersion: Int): NotarisationHandler {
         return ValidatingNotaryFlow(otherParty, this)
     }
