+++ node/src/main/kotlin/net/corda/node/services/transactions/SimpleNotaryService.kt
@@ -4,7 +4,7 @@ import net.corda.core.identity.Party
 import net.corda.core.node.services.TrustedAuthorityNotaryService
 import net.corda.core.node.services.ServiceType
 import net.corda.core.node.services.TimeWindowChecker
-import net.corda.flows.NotaryFlow
+import net.corda.flows.NotarisationHandler
 import net.corda.node.services.api.ServiceHubInternal
 
 /** A simple Notary service that does not perform transaction validation */
@@ -16,7 +16,7 @@ class SimpleNotaryService(override val services: ServiceHubInternal) : TrustedAu
     override val timeWindowChecker = TimeWindowChecker(services.clock)
     override val uniquenessProvider = PersistentUniquenessProvider()
 
-    override fun createServiceFlow(otherParty: Party, platformVersion: Int): NotaryFlow.Service {
+    override fun createHandler(otherParty: Party, platformVersion: Int): NotarisationHandler {
         return NonValidatingNotaryFlow(otherParty, this)
     }
