+++ node/src/main/kotlin/net/corda/node/services/transactions/ValidatingNotaryService.kt
@@ -4,7 +4,7 @@ import net.corda.core.identity.Party
 import net.corda.core.node.services.TrustedAuthorityNotaryService
 import net.corda.core.node.services.ServiceType
 import net.corda.core.node.services.TimeWindowChecker
-import net.corda.flows.NotaryFlow
+import net.corda.flows.NotarisationHandler
 import net.corda.node.services.api.ServiceHubInternal
 
 /** A Notary service that validates the transaction chain of the submitted transaction before committing it */
@@ -16,7 +16,7 @@ class ValidatingNotaryService(override val services: ServiceHubInternal) : Trust
     override val timeWindowChecker = TimeWindowChecker(services.clock)
     override val uniquenessProvider = PersistentUniquenessProvider()
 
-    override fun createServiceFlow(otherParty: Party, platformVersion: Int): NotaryFlow.Service {
+    override fun createHandler(otherParty: Party, platformVersion: Int): NotarisationHandler {
         return ValidatingNotaryFlow(otherParty, this)
     }
