+++ samples/notary-demo/src/main/kotlin/net/corda/notarydemo/flows/RPCStartableNotarisationFlow.kt
@@ -2,7 +2,7 @@ package net.corda.notarydemo.flows
 
 import net.corda.core.flows.StartableByRPC
 import net.corda.core.transactions.SignedTransaction
-import net.corda.flows.NotaryFlow
+import net.corda.flows.NotarisationFlow
 
 @StartableByRPC
-class RPCStartableNotaryFlowClient(stx: SignedTransaction) : NotaryFlow.Client(stx)
+class RPCStartableNotarisationFlow(stx: SignedTransaction) : NotarisationFlow(stx)
